requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Marshmallow
Flask-Marshmallow
Marshmallow-SQLAlchemy
Flask-Babel
python-dotenv
Flask-Uploads
Flask-JWT-Extended
Flask-Authorize
Requests
uwsgi